\name{baseball}
\alias{baseball}
\docType{data}
\title{baseball data}
\description{
  
}
\usage{data(baseball)}
\format{

  \describe{
    \name{baseball}
    \item{Player}{Player} 
    \item{hat.P}{the batting average of  player after their first 45 times at bat during the 1970 season}
    \item{x1}{career batting average through the end of 1969 season}
    \item{x2}{x2}{number of times at bat until 1969 season}
    \item{P}{known batting average for the entire 1970 season (true value)}
  }
}
\source{



}
\examples{
  data(baseball)
}
\keyword{datasets}
