% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{body}
\alias{body}
\title{Return the middle n rows and columns of an object}
\usage{
body(data, nrows = 15, ncols = 10, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Return the middle n rows and columns of an object
}
