% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{ce}
\alias{ce}
\title{Combine character values into a vector or list without quotes}
\usage{
ce(...)
}
\arguments{
\item{...}{}
}
\description{
Combine character values into a vector or list without quotes
}
\examples{
ce(a,b,c)
}
