% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{clusterData}
\alias{clusterData}
\title{Hierarchical clustering of data}
\usage{
clusterData(
  data,
  method = "hclust",
  rows = NULL,
  cols = NULL,
  inf = NULL,
  na = NULL,
  ...
)
}
\arguments{
\item{data}{matrix or dataframe}

\item{method}{hclust}

\item{rows}{cluster rows}

\item{cols}{cluster columns}

\item{inf}{handling of non-finite values}

\item{na}{handling of missing values}

\item{...}{}
}
\value{
clust
}
\description{
Hierarchical clustering of data
}
