% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{collapse}
\alias{collapse}
\title{Collapse duplicate ids}
\usage{
collapse(
  data,
  ids,
  select_by = NULL,
  average_by = NULL,
  decreasing = TRUE,
  ...
)
}
\arguments{
\item{data}{matrix/data.frame}

\item{ids}{vector of row identifiers}

\item{select_by}{FUN for calculating stat by which duplicates are selected (applied to each row)}

\item{average_by}{FUN for calculating averages of duplicate rows}

\item{decreasing}{sort selecting stat increasingly/decreasingly (default = TRUE)}

\item{...}{}
}
\value{

}
\description{
Collapse duplicate ids
}
