% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{compare}
\alias{compare}
\title{Colored printing to compare two dataframes}
\usage{
compare(m1, m2, ncol = 10, digits = 3, sep = "/", width = 20)
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Colored printing to compare two dataframes
}
