% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{compareDF}
\alias{compareDF}
\title{Colored printing to compare two dataframes}
\usage{
compareDF(
  df1,
  df2,
  ncol = 10,
  signif = 3,
  sep = "/",
  marker = "*",
  color = "red"
)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Colored printing to compare two dataframes
}
