% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{convertGeneIDs}
\alias{convertGeneIDs}
\title{Gene ID conversion}
\usage{
convertGeneIDs(
  ids,
  from = "ENTREZID",
  to = "SYMBOL",
  annotation = org.Hs.eg.db::org.Hs.eg.db,
  multiVals = "collapse",
  ...
)
}
\arguments{
\item{multiVals}{first, collapse}

\item{...}{}
}
\description{
Gene ID conversion
}
\examples{
convertGeneIDs(c("3098", "3099"))
}
