% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_functions.R
\name{convertGeneSets}
\alias{convertGeneSets}
\title{Convert gene sets between different formats}
\usage{
convertGeneSets(genesets, from = NULL, to = "list", term = term, gene = gene)
}
\arguments{
\item{gene}{}
}
\description{
Convert gene sets between different formats
}
