% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_functions.R
\name{convertGeneSpecies}
\alias{convertGeneSpecies}
\title{Convert gene IDs between species}
\usage{
convertGeneSpecies(
  genes,
  taxon = "10090",
  direction = "from",
  id_type = "symbol",
  na.omit = TRUE
)
}
\arguments{
\item{na.omit}{}
}
\description{
Convert gene IDs between species
}
\examples{
convertGeneSpecies("Egfr", taxon = "10090")
convertGeneSpecies("EGFR", taxon = "10090", direction = "to")
}
