% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{correl}
\alias{correl}
\title{Calculate correlations between any types of variables}
\usage{
correl(x, y, method_numeric = "spearman", na.omit = TRUE, ...)
}
\arguments{
\item{na.omit}{}
}
\description{
Calculate correlations between any types of variables
}
\examples{
correl(1:10, sample(1:10))
correl(rep(LETTERS, 5), seq(rep(LETTERS, 5)))
correl(rep(LETTERS, 10), sample(rep(LETTERS, 10)))
}
