% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{COX}
\alias{COX}
\title{Cox Proportional Hazards Analysis}
\usage{
COX(
  formula,
  data = NULL,
  update = NULL,
  surv = NULL,
  robust = FALSE,
  response = "surv",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Cox Proportional Hazards Analysis
}
