% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{cutstr}
\alias{cutstr}
\title{Limit number of characters in a string}
\usage{
cutstr(x, maxchar = 25, add = "...", add_incl = TRUE)
}
\arguments{
\item{x}{character string}

\item{maxchar}{max. number of characters}

\item{add}{string added to truncated objects}

\item{add_incl}{include added string in maxchar}
}
\value{

}
\description{
Limit number of characters in a string
}
