% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cxheatmap}
\alias{cxheatmap}
\title{Heatmap plotting with ComplexHeatmap}
\usage{
cxheatmap(
  data,
  rowdf = NULL,
  coldf = NULL,
  scale = FALSE,
  cluster_rows = NULL,
  cluster_cols = NULL,
  rowdf_side = "left",
  coldf_side = "top",
  rowdf_legend = TRUE,
  coldf_legend = TRUE,
  legend_border = "black",
  fontsize = 12,
  rowcex = NULL,
  colcex = 1,
  heatpal = NULL,
  border = NULL,
  title = NULL,
  colors = NULL,
  inf = F,
  na = 0,
  mat = NULL,
  markoob = FALSE,
  markshape = 4,
  marksize = NULL,
  na_col = "grey",
  maxchar = 35,
  ...
)
}
\arguments{
\item{data}{matrix or dataframe}

\item{rowdf}{dataframe with row annotations}

\item{coldf}{dataframe with column annotations}

\item{scale}{z-scale 'rows' or 'cols'}

\item{cluster_rows}{cluster rows}

\item{cluster_cols}{cluster columns}

\item{rowdf_side}{left/right}

\item{coldf_side}{top/bottom}

\item{rowdf_legend}{show rowdf legend}

\item{coldf_legend}{show coldf legend}

\item{legend_border}{border of legends}

\item{fontsize}{base fontsize}

\item{rowcex}{sizefactor for row annotations}

\item{colcex}{sizefactor for col annotations}

\item{heatpal}{color palette ('circlize::colorRamp2')}

\item{border}{cell border}

\item{title}{title of colorscale legend}

\item{colors}{list of colors for annotations}

\item{inf}{handling of Inf values for clustering}

\item{na}{handling of NA values for clustering}

\item{mat}{logical matrix indicating where to draw marks}

\item{markoob}{mark out-of-bounds values}

\item{markshape}{shape of cell marks}

\item{marksize}{size of cell marks}

\item{na_col}{color of NA values}

\item{maxchar}{max. length of colnames/rownames}

\item{...}{}
}
\value{

}
\description{
Heatmap plotting with ComplexHeatmap
}
\examples{
cxheatmap(rmat(50, 30), coldf = data.frame(row.names = colnames(rmat(50, 30)), group = sample(LETTERS[1:5], size = 30, replace = T)))
cxheatmap(rmat(50, 30), border = c("black", 2))

}
\seealso{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
