% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{dbugapply}
\alias{dbugapply}
\title{Drop-in replacement for lapply to quickly identify failing items}
\usage{
dbugapply(X, FUN, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Drop-in replacement for lapply to quickly identify failing items
}
