% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{dedupl}
\alias{dedupl}
\title{Rename duplicated strings}
\usage{
dedupl(x, sep = ".", index = NULL, ...)
}
\arguments{
\item{x}{character vector}

\item{sep}{separator}

\item{index}{indices added to duplicated elements}

\item{...}{arguments passed to 'duplicated'}
}
\value{

}
\description{
Rename duplicated strings
}
\examples{
replicate(3, paste(LETTERS[1:3], collapse = "")) \%>\% dedupl()
replicate(3, paste(LETTERS[1:3], collapse = "")) \%>\% dedupl(sep = "_", index = letters)

}
