% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{duplicates}
\alias{duplicates}
\title{Get non-unique elements}
\usage{
duplicates(x, index = FALSE)
}
\arguments{
\item{x}{vector}

\item{index}{return as index}
}
\description{
Get non-unique elements
}
\examples{
duplicates(c("A", "A", "B"))
}
