% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq_functions.R
\name{find_markers}
\alias{find_markers}
\title{Find markers for groups in a Seurat object}
\usage{
find_markers(
  seu,
  group_by = NULL,
  group_oi = NULL,
  group_ref = NULL,
  assay = NULL,
  slot = "data",
  logfc_thresh = 0.1,
  min_pct = 0.1,
  padj_thresh = 0.05,
  test_use = "wilcox",
  rename = TRUE,
  n_cores = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Find markers for groups in a Seurat object
}
\examples{
get_seu_dataset() |> find_markers(group_by = "seurat_annotations", logfc_thresh = 1, min_pct = 0.5, padj_thresh = 0.05)
get_seu_dataset() |> find_markers(group_by = "seurat_annotations", group_oi = "DC", group_ref = "NK")
}
