% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{genPalettes}
\alias{genPalettes}
\title{Generate distinct color palettes}
\usage{
genPalettes(
  n = 1,
  length_each = 3,
  dist = 0.75,
  saturation = c(0.5, 0.7),
  lightness = c(0.5, 0.7),
  cvd = "protan",
  cvd_severity = 0.2,
  seed = 123,
  ...
)
}
\arguments{
\item{n}{Number of palettes}

\item{length_each}{Length of each palette (either a single digit or a vector of length n)}

\item{dist}{Distance between palettes}

\item{...}{}
}
\description{
Generate distinct color palettes
}
\examples{
genPalettes(n = 10) \%L>\% showColors()
}
