% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq_functions.R
\name{get_seu_dataset}
\alias{get_seu_dataset}
\title{Load Seurat datasets}
\usage{
get_seu_dataset(ds = "pbmc3k")
}
\arguments{
\item{ds}{Dataset name}
}
\description{
Load Seurat datasets
}
\examples{
get_seu_dataset()
}
