% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq_functions.R
\name{get_top_markers}
\alias{get_top_markers}
\title{Extract top markers from find_markers() results}
\usage{
get_top_markers(
  markers,
  n = 5,
  per_group = TRUE,
  distinct = FALSE,
  group_col = "group",
  feature_col = "gene",
  filter_pattern = NULL,
  ...
)
}
\arguments{
\item{markers}{Dataframe}

\item{n}{Number of top features}

\item{per_group}{Get features per group}

\item{distinct}{Count each feature only once}

\item{...}{}
}
\description{
Extract top markers from find_markers() results
}
\examples{
get_seu_dataset() |> find_markers(group_by = "seurat_annotations") |> get_top_markers()
}
