% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_functions.R
\name{getBiomaRt}
\alias{getBiomaRt}
\title{Retrieve data from BiomaRt}
\usage{
getBiomaRt(
  genes,
  idtype = "hgnc_symbol",
  attributes = "transcript_biotype",
  biotype = "protein_coding",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Retrieve data from BiomaRt
}
