% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{getColors}
\alias{getColors}
\title{Generate color mappings for discrete data columns in a dataframe}
\usage{
getColors(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{

}
\description{
Generate color mappings for discrete data columns in a dataframe
}
