% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{getGeneSets}
\alias{getGeneSets}
\title{Get gene sets from msigdbr}
\usage{
getGeneSets(
  collections = c("H", "C2|CP:KEGG"),
  species = "human",
  id_type = "gene_symbol",
  format = c("list", "dataframe", "GeneSetCollection"),
  ...
)
}
\arguments{
\item{collections}{Collections (msigdbr::msigdbr_collections()) and subcollections}

\item{format}{}
}
\description{
Get gene sets from msigdbr
}
\examples{
getGeneSets(c("H", "C2|CP:KEGG|CP:REACTOME"))
}
