% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{getGEOdata}
\alias{getGEOdata}
\title{Get data from the GEO database}
\usage{
getGEOdata(
  ID = "GSE222693",
  geodir = "~/myScratch/GEO",
  fetch_files = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Get data from the GEO database
}
\examples{
GSE234099 <- getGEOdata(ID = "GSE234099", fetch_files = FALSE)
GSE222693 <- getGEOdata(ID = "GSE222693")
}
