% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_functions.R
\name{getGOgenes}
\alias{getGOgenes}
\title{Get GO gene sets}
\usage{
getGOgenes(
  db = org.Hs.eg.db::org.Hs.eg.db,
  keytype = "SYMBOL",
  ontology = c("BP", "MF", "CC"),
  minsize = 3
)
}
\arguments{
\item{minsize}{}
}
\description{
Get GO gene sets
}
