% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_functions.R
\name{getMetabolicAtlas}
\alias{getMetabolicAtlas}
\title{Get gene/metabolite sets from metabolicatlas}
\usage{
getMetabolicAtlas(type = "genes")
}
\arguments{
\item{type}{}
}
\description{
Get gene/metabolite sets from metabolicatlas
}
