% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_functions.R
\name{getMitoCarta}
\alias{getMitoCarta}
\title{Get gene sets from MitoCarta3}
\usage{
getMitoCarta(file = "Human.MitoCarta3.0.xls")
}
\arguments{
\item{file}{}
}
\description{
Get gene sets from MitoCarta3
}
