% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{getPaletteer}
\alias{getPaletteer}
\title{Select colors using the 'paletteer' package}
\usage{
getPaletteer(length = NULL, continuous = TRUE, discrete = TRUE)
}
\arguments{
\item{length}{Integer/vector of desired length(s)}

\item{continuous}{TRUE}

\item{discrete}{TRUE}
}
\description{
Select colors using the 'paletteer' package
}
\examples{
getPaletteer(length = 2)$colors \%>\% showColors()
}
