% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{getRanks}
\alias{getRanks}
\title{Get feature ranks from differential abundance testing}
\usage{
getRanks(data, rank_by = 1, type = NULL)
}
\arguments{
\item{type}{}
}
\description{
Get feature ranks from differential abundance testing
}
