% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggboxplot}
\alias{ggboxplot}
\title{Boxplot with individual data points}
\usage{
ggboxplot(
  df,
  x,
  y,
  colour = "grey30",
  ymax = NULL,
  point_size = 2.5,
  fontsize = 16,
  linewidth = NULL,
  width = 0.8,
  varwidth = FALSE,
  staplewidth = 0.4,
  y_at_zero = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Boxplot with individual data points
}
\examples{
iris |> ggboxplot(Species, Sepal.Length, colour=Petal.Length)
iris |> ggboxplot(Species, Sepal.Length, colour=Petal.Length) |> ggpairwise()
}
