% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggdesign}
\alias{ggdesign}
\title{Plot experimental design matrix}
\usage{
ggdesign(
  design,
  columns = NULL,
  colors = NULL,
  label = NULL,
  legend = NULL,
  fontsize = 15,
  nacol = "grey70",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot experimental design matrix
}
