% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggpairwise}
\alias{ggpairwise}
\title{Add pairwise comparisons to ggboxplot}
\usage{
ggpairwise(
  gg,
  padj_thres = 0.05,
  fontsize = 10,
  max_add = NULL,
  spacing = NULL,
  label_offset = 0.02,
  linewidth = NULL,
  linetype = "solid",
  q = 0.99,
  y_at_zero = TRUE,
  ymax = NULL,
  na.omit = FALSE,
  test = c("wilcox.test", "t.test"),
  TESTFUN = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Add pairwise comparisons to ggboxplot
}
\examples{
iris |> ggboxplot(Species, Sepal.Length, colour=Petal.Length) |> ggpairwise()
}
