% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggpca}
\alias{ggpca}
\title{Principal component analysis and plotting of data}
\usage{
ggpca(
  data,
  design = NULL,
  mapping = aes(),
  center = TRUE,
  scale = TRUE,
  na = NULL,
  n = NULL,
  label = FALSE,
  digits = 2,
  colors = NULL,
  size = 3.5,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Quick plotting of PCA results.
For advanced analysis, see e.g. browseVignettes("PCAtools").
}
