% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggpdf}
\alias{ggpdf}
\title{Save ggplot as pdf}
\usage{
ggpdf(gg, file, width = 3000, height = 2500, dpi = 300, units = "px", ...)
}
\arguments{
\item{...}{}
}
\description{
Save ggplot as pdf
}
