% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggpng}
\alias{ggpng}
\title{Save ggplot as png}
\usage{
ggpng(gg, file, width = 3000, height = 2500, dpi = 300, units = "px", ...)
}
\arguments{
\item{...}{}
}
\description{
Save ggplot as png
}
