% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggseabar}
\alias{ggseabar}
\title{GSEA bar plot}
\usage{
ggseabar(
  results,
  x = NULL,
  y = NULL,
  label = NULL,
  sort_by = NULL,
  sort_abs = FALSE,
  top_by = NULL,
  top_n = 20,
  top_n_up = NULL,
  top_n_down = NULL,
  labsize = 12,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GSEA bar plot
}
