% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggseadot}
\alias{ggseadot}
\title{GSEA dotplot}
\usage{
ggseadot(
  data,
  x = NES,
  y = term,
  colour = -log10(padj),
  size = n_enriched_genes,
  title = NULL,
  sort = TRUE,
  top_n_up = 15,
  top_n_down = 15,
  xmin = NULL,
  xmax = NULL,
  fontsize = 12,
  size_limits = c(0, NA),
  size_range = c(0.01, 5),
  maxchar = NULL,
  margins = ggplot2::margin(t = 8, r = 8, l = 8, b = 8)
)

ggseadot(
  data,
  x = NES,
  y = term,
  colour = -log10(padj),
  size = n_enriched_genes,
  title = NULL,
  sort = TRUE,
  top_n_up = 15,
  top_n_down = 15,
  xmin = NULL,
  xmax = NULL,
  fontsize = 12,
  size_limits = c(0, NA),
  size_range = c(0.01, 5),
  maxchar = NULL,
  margins = ggplot2::margin(t = 8, r = 8, l = 8, b = 8)
)
}
\arguments{
\item{margins}{}

\item{...}{}
}
\description{
GSEA dotplot

Dotplot of GSEA results
}
\examples{
data.frame(term=rnames(100), NES=runif(100,-3,3), padj=runif(100)) |> ggseadot(size=NULL)
data.frame(term=rnames(100), NES=runif(100,-3,3), padj=runif(100)) |> ggseadot(x=log10(padj), colour=NES, size=NULL)
}
