% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggtopfeatures}
\alias{ggtopfeatures}
\title{Barplot of the top features of an expression matrix}
\usage{
ggtopfeatures(
  data,
  rel = FALSE,
  topn = 10,
  col.high = "#e63c3c",
  col.low = "#f0c3bb",
  col.other = "#dbdbdb",
  fontsize = 18,
  label.cex = 0.6,
  label.rot = 0,
  x.rot = 90,
  width = 0.9,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Barplot of the top features of an expression matrix
}
\examples{
rmat(nrow = 100) |> ggtopfeatures()
}
