% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{Volcano plots using ggplot2}
\usage{
ggvolcano(
  data,
  x = NULL,
  y = NULL,
  color = NULL,
  label = NULL,
  shape = NULL,
  stroke = NA,
  nlabels = NULL,
  fontsize = 16,
  labface = "plain",
  lab_size = 12,
  repel = 2,
  attract = NULL,
  box.padding = 0.5,
  max_overlaps = Inf,
  seed = 123,
  ptres = 0.05,
  clip = FALSE,
  symlim = TRUE,
  expand = c(0, 0),
  nbreaks_x = 7,
  nbreaks_y = 7,
  xlim = NULL,
  ylim = NULL,
  color_up = "#eb9d0e",
  color_down = "#146bc7",
  color_nonsig = "#4d4d4d",
  label_up = "up",
  label_down = "down",
  label_nonsig = "not signif.",
  autolabel = NULL,
  show_nonsig = TRUE,
  segment.alpha = 0.6,
  max.time = 30,
  max.iter = 10^6,
  show_grid = TRUE,
  title = NULL,
  title_size = NULL,
  point_size = 2,
  scale_size = FALSE,
  axis_size = NULL,
  leg_size = NULL,
  leg_key_size = 4,
  lwd = 0.8,
  at_zero = FALSE,
  clip_frame = "off",
  ...
)
}
\arguments{
\item{data}{Dataframe}

\item{x}{Variable to plot on x-axis}

\item{y}{Variable to plot on y-axis}

\item{...}{}
}
\description{
Volcano plots using ggplot2
}
\examples{
data.frame(row.names = LETTERS, lfc = runif(length(LETTERS), -3, 3), padj = runif(length(LETTERS))) |> ggvolcano()
}
