% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{gradient}
\alias{gradient}
\title{Get color gradients}
\usage{
gradient(colors, breaks = NULL, type = c("ggplot2", "colorRamp2"), name = NULL)
}
\arguments{
\item{colors}{vector of colors}

\item{breaks}{vector of breaks (required for 'colorRamp2')}

\item{type}{generate gradient as ggplot2, colorRamp2 or none}

\item{name}{use a pre-defined gradient}
}
\description{
Get color gradients
}
\examples{
gradient(colors = c("-3 "= "blue", "0" = "white", "3" = "orangered"))
gradient(name = "bluered3", type = "colorramp2")
ggplot(iris, aes(Sepal.Length, Petal.Length, color = Sepal.Width)) + geom_point() + gradient(c("red", "green", "blue"), breaks = c(0,1,5))
ggplot(iris, aes(Sepal.Length, Petal.Length, color = Sepal.Width)) + geom_point() + gradient(c("red", "green", "blue"))
ComplexHeatmap::Heatmap(iris[,-5], col = gradient(name = "blueyellow3", type = "colorramp2", breaks = c(0,4,8)))
}
