% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{gseaplot}
\alias{gseaplot}
\title{GSEA plot modified from 'enrichplot::gseaplot2'}
\usage{
gseaplot(gsearesults, n = 3)
}
\arguments{
\item{n}{}
}
\description{
GSEA plot modified from 'enrichplot::gseaplot2'
}
