% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{head2}
\alias{head2}
\title{Return the first n rows and columns of an object}
\usage{
head2(data, nrows = 15, ncols = 10, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Return the first n rows and columns of an object
}
