% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{intercalate}
\alias{intercalate}
\title{Intercalate one vector with another}
\usage{
intercalate(x, y, end = TRUE)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector/value 2}

\item{end}{add value at the end}
}
\description{
Intercalate one vector with another
}
\examples{
intercalate(1:5, 0)
}
