% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{intersection}
\alias{intersection}
\title{Intersection of all elements}
\usage{
intersection(...)
}
\arguments{
\item{...}{}
}
\description{
Intersection of all elements
}
\examples{
intersection(LETTERS[1:10], LETTERS[5:8], LETTERS[7:12])
}
