% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{lgrep}
\alias{lgrep}
\title{Pattern Matching and Replacement}
\usage{
lgrep(pattern, x, ...)
}
\arguments{
\item{pattern}{Multiple patterns}

\item{...}{}
}
\description{
Pattern Matching and Replacement
}
