% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listpipes.R
\name{listpipe}
\alias{listpipe}
\title{List pipe function}
\usage{
listpipe(lhs, rhs, env, parallel = FALSE, simplify = FALSE, qsub = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
List pipe function
}
