% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{make_filenames}
\alias{make_filenames}
\title{Safe filenames}
\usage{
make_filenames(
  names,
  replacement = "_",
  unique = TRUE,
  sep = .Platform$file.sep,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Safe filenames
}
\examples{
make_filenames("file/na.me")
}
