% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{matScale}
\alias{matScale}
\title{Scale rows (columns) of a matrix}
\usage{
matScale(data, rows = FALSE, cols = FALSE, FUN = scale, ...)
}
\arguments{
\item{data}{Matrix or data.frame}

\item{rows}{Scale rows (TRUE/FALSE)}

\item{cols}{Scale columns (TRUE/FALSE)}

\item{FUN}{Function used for scaling}

\item{...}{}
}
\value{

}
\description{
Apply FUN ('scale' by default) to the rows or columns of a numeric matrix.
}
\examples{
matScale(rmat(5, 5), rows = TRUE)
}
