% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{na_skip}
\alias{na_skip}
\title{Skip rows/columns containing NA values}
\usage{
na_skip(data, skip = "rows")
}
\arguments{
\item{data}{Dataset}

\item{skip}{Skip rows (columns) containing NA values.}
}
\description{
Function for fast subsetting of data (see also 'na.omit()').
}
