% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{naSkip}
\alias{naSkip}
\title{Skip rows/columns containing NA values}
\usage{
naSkip(data, skip = "rows")
}
\arguments{
\item{data}{Dataset}

\item{skip}{Skip rows (columns) containing NA values.}
}
\value{

}
\description{
Function for fast subsetting of data (see also 'na.omit()').
}
