% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_getSamplesheet}
\alias{nf_getSamplesheet}
\title{Import samplesheet from nf-core/rnaseq pipeline}
\usage{
nf_getSamplesheet(file, exclude = c("fastq_1", "fastq_2", "strandedness"), ...)
}
\arguments{
\item{...}{}
}
\description{
Import samplesheet from nf-core/rnaseq pipeline
}
