% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_getVersion}
\alias{nf_getVersion}
\title{Get version of the used nf-core/rnaseq pipeline}
\usage{
nf_getVersion(ymlfile)
}
\arguments{
\item{ymlfile}{}
}
\value{

}
\description{
Get version of the used nf-core/rnaseq pipeline
}
