% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_importTX}
\alias{nf_importTX}
\title{Import counts from nf-core/rnaseq pipeline}
\usage{
nf_importTX(nfdir, ...)
}
\arguments{
\item{nfdir}{Results directory}

\item{...}{}
}
\value{

}
\description{
Import counts from nf-core/rnaseq pipeline
}
