% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_preprocessing}
\alias{nf_preprocessing}
\title{Import and plot preprocessing statistics from nf-core/rnaseq pipeline}
\usage{
nf_preprocessing(multiqc_dir, design = NULL, ignore = FALSE)
}
\arguments{
\item{ignore}{}
}
\value{

}
\description{
Import and plot preprocessing statistics from nf-core/rnaseq pipeline
}
