% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_summary}
\alias{nf_summary}
\title{Import and plot preprocessing statistics from nf-core/rnaseq pipeline}
\usage{
nf_summary(nfdir, design = NULL, ignore = FALSE)
}
\arguments{
\item{ignore}{}
}
\description{
Import and plot preprocessing statistics from nf-core/rnaseq pipeline
}
