% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_tximport}
\alias{nf_tximport}
\title{Import counts from nf-core/rnaseq pipeline using tximport}
\usage{
nf_tximport(
  nfdir,
  samplesheet = NULL,
  subdir = "star_salmon",
  pattern = "quant.sf",
  tx2gene = "tx2gene.tsv",
  type = "salmon",
  gene = "SYMBOL",
  ...
)
}
\arguments{
\item{nfdir}{results directory}

\item{samplesheet}{samplesheet data frame}

\item{subdir}{star_salmon}

\item{pattern}{quant.sf (salmon output)}

\item{tx2gene}{tx2gene.tsv (salmon output)}

\item{...}{}
}
\description{
Import counts from nf-core/rnaseq pipeline using tximport
}
