% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{nf_versions}
\alias{nf_versions}
\title{Get versions of tools used in the nf-core/rnaseq pipeline}
\usage{
nf_versions(ymlfile)
}
\arguments{
\item{ymlfile}{}
}
\description{
Get versions of tools used in the nf-core/rnaseq pipeline
}
