% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{padjust}
\alias{padjust}
\title{Adjust P-values for Multiple Comparisons}
\usage{
padjust(p, method = "fdr", ...)
}
\arguments{
\item{p}{vector or matrix of p-values}

\item{...}{}
}
\description{
Adjust a vector or matrix of p-values derived from the same data for multiple testing.
}
