% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{partition}
\alias{partition}
\title{Divide vector into n groups}
\usage{
partition(x, n = 3)
}
\arguments{
\item{n}{}
}
\description{
Divide vector into n groups
}
\examples{
partition(LETTERS[1:10], 3)
}
