% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{pval_format}
\alias{pval_format}
\title{Format p-values for printing}
\usage{
pval_format(p, min = 0.001, scientific = NULL, add = "p ", stars = FALSE)
}
\arguments{
\item{stars}{}
}
\description{
Format p-values for printing
}
