% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listpipes.R
\name{qapply}
\alias{qapply}
\title{SGE qsub wrapper function}
\usage{
qapply(
  X,
  FUN,
  keep_files = FALSE,
  max_cores = 30,
  jobname = NULL,
  tmpdir = NULL,
  env = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
SGE qsub wrapper function
}
