% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{rangescale}
\alias{rangescale}
\title{Scale data within range}
\usage{
rangescale(data, from = 0, to = 1)
}
\arguments{
\item{data}{Matrix or vector}

\item{from}{minimum}

\item{to}{maximum}
}
\description{
Scale data within range
}
\examples{
rangescale(1:5)
}
