\name{rats}
\alias{rats}
\docType{data}
\title{Rat Growth Data}
\description{
Data on the weights of 30 rats each week for 5 weeks.
}
\usage{data(rats)}
\format{
  A data frame with 150 observations on the following 3 variables.
  \describe{
    \item{rat}{a factor with levels 1-30}
    \item{x}{age in days}
    \item{y}{rat weight (units unspecified)}
  }
}

\source{
Gelfand, A. E., Hills, S. E., Racine-Poon, A. and Smith, A. F. M. (1990)
Illustration of Bayesian inference in normal data models using Gibbs sampling.
\emph{Journal of the American Statistical Association}, \bold{85}, 972--985.
}
\examples{
data(rats)
}
\keyword{datasets}
