% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{rdataframe}
\alias{rdataframe}
\title{Generate a random data.frame}
\usage{
rdataframe(nrow = 5, ncol = 5)
}
\arguments{
\item{ncol}{}
}
\description{
Generate a random data.frame
}
\examples{
rdataframe()
}
