% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readTables}
\alias{readTables}
\title{Read all sheets from .xlsx file}
\usage{
readTables(file, rowNames = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Read all sheets from .xlsx file
}
\examples{
writeTables(list(mtcars = mtcars, iris = iris), file = "example.xlsx")
readTables(file = "example.xlsx")
}
