% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{rlist}
\alias{rlist}
\title{Generate a random list}
\usage{
rlist(length = 5, items = 1:3, space = LETTERS)
}
\arguments{
\item{length}{length of the list}

\item{items}{(variable) number of items per list entry}

\item{space}{set of items to sample}
}
\value{
list
}
\description{
Generate a random list
}
\examples{
rlist()
rlist(length = 3, items = 2:4, space = 1:10)
}
