% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{rmat}
\alias{rmat}
\title{Generate a matrix of random numbers}
\usage{
rmat(nrow = 3, ncol = 5, FUN = runif, ...)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{FUN}{random number generator}

\item{...}{parameters other than n passed to FUN}
}
\value{

}
\description{
Returns a matrix of size nrow x ncol from random numbers generated by FUN.
}
\seealso{
\code{\link[stats]{runif}}, \code{\link[stats]{rnorm}}, \code{\link[stats]{rexp}}
}
