% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{roundup}
\alias{roundup}
\title{Rounding of numbers}
\usage{
roundup(x, digits = 0, ...)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of decimal places}

\item{...}{other arguments passed to ceiling/floor}
}
\description{
Round up
}
