% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runfGSEA}
\alias{runfGSEA}
\title{GSEA with fGSEA}
\usage{
runfGSEA(
  data,
  genesets,
  rank_by = c(stat, baseMean, svalue),
  type = c("", "+", "p"),
  seed = 123
)
}
\arguments{
\item{seed}{}
}
\description{
GSEA with fGSEA
}
