% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runGSEA}
\alias{runGSEA}
\title{GSEA with clusterProfiler}
\usage{
runGSEA(
  data,
  genesets = NULL,
  rank_by = c(stat, baseMean, svalue),
  type = c("", "+", "p"),
  as.df = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GSEA with clusterProfiler
}
