% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runGSVA}
\alias{runGSVA}
\title{Single-sample gene set enrichment with GSVA}
\usage{
runGSVA(
  data,
  genesets = NULL,
  method = "gsva",
  kcdf = "Gaussian",
  ncores = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Single-sample gene set enrichment with GSVA
}
