% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runLIMMA}
\alias{runLIMMA}
\title{Differential expression analysis with LIMMA}
\usage{
runLIMMA(
  data,
  design,
  formula = ~1,
  contrasts = NULL,
  trend = TRUE,
  robust = FALSE,
  p.adj.method = "fdr",
  normalize = FALSE,
  norm.method = NULL,
  ...
)
}
\arguments{
\item{data}{Data}

\item{design}{Experimental design/colData}

\item{formula}{Formula}

\item{contrasts}{Named list of contrasts, specified as c(factor, level, reflevel)}

\item{trend}{Trend}

\item{robust}{Robust}

\item{p.adj.method}{P.adj.method}

\item{normalize}{Normalize}

\item{norm.method}{Normalization method}

\item{...}{}
}
\description{
Differential expression analysis with LIMMA
}
