% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runORA}
\alias{runORA}
\title{Gene set over-representation analysis}
\usage{
runORA(genes, genesets = NULL, p_cutoff = 0.25, ...)
}
\arguments{
\item{genes}{vector of genes}

\item{genesets}{gene set dataframe}

\item{...}{}
}
\description{
Gene set over-representation analysis
}
\examples{
getMSigDB("H")$HALLMARK_GLYCOLYSIS |> runORA()
}
