% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{saveplot}
\alias{saveplot}
\title{Save plot to file}
\usage{
saveplot(
  p,
  file = NULL,
  dev = "png",
  width = 3000,
  height = 2500,
  dpi = 300,
  units = "px",
  bg = "white",
  ggsave = TRUE,
  PRINTFUN = print,
  ...
)
}
\arguments{
\item{p}{plot}

\item{file}{filename}

\item{dev}{device}

\item{...}{}
}
\description{
Save plot to file
}
