% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq_functions.R
\name{seu_dotplot}
\alias{seu_dotplot}
\title{DotPlot wrapper for Seurat}
\usage{
seu_dotplot(
  seu,
  group_by = NULL,
  features = NULL,
  dot_scale = NULL,
  col_mid = NULL,
  scale = FALSE,
  col_low = "grey98",
  col_high = "#0000b5",
  col_min = NA,
  col_max = NA,
  fontsize = 12,
  lwd = 0.5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
DotPlot wrapper for Seurat
}
\examples{
get_seu_dataset() |> seu_dotplot(group_by = "seurat_annotations", features = c( "S100A9", "S100A8", "LGALS2", "CD79A", "GZMB", "HLA-DQA1", "CCL5", "NKG7"))
}
