% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq_functions.R
\name{seu_downsample}
\alias{seu_downsample}
\title{Downsample Seurat object}
\usage{
seu_downsample(
  seu,
  group_by = "celltype",
  n_cells = NULL,
  min_cells = NULL,
  seed = 123
)
}
\arguments{
\item{seu}{Seurat object}

\item{group_by}{Cluster}

\item{n_cells}{Number of target cell}

\item{min_cells}{Minimum number of cells per group}

\item{seed}{Random seed}
}
\description{
Downsample Seurat object
}
\examples{
get_seu_dataset() |> seu_downsample(group_by = "seurat_annotations", n_cells = 100, min_cells = 10)
}
