% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq_functions.R
\name{seu_umap}
\alias{seu_umap}
\title{UMAP plot wrapper for Seurat}
\usage{
seu_umap(
  seu,
  features = NULL,
  groups = NULL,
  label = FALSE,
  alpha = 1,
  assay = NULL,
  slot = NULL,
  colors = NULL,
  min_cutoff = "q05",
  max_cutoff = "q95",
  reduction = "umap",
  pt_shape = 16,
  pt_size = 0.5,
  title_size = 18,
  axis_size = 16,
  axis_text_size = 14,
  legend_key_size = 18,
  legend_text_size = 18,
  lwd = 0.5,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
UMAP plot wrapper for Seurat
}
\examples{
seu <- get_seu_dataset() |> ScaleData() |> FindVariableFeatures() |> RunPCA() |> RunUMAP(dims = 1:10)
seu_umap(seu, features = "S100A9")
seu_umap(seu, groups = "seurat_annotations")
}
