% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{showColors}
\alias{showColors}
\title{Display colors in console}
\usage{
showColors(colors)
}
\arguments{
\item{colors}{input colors}
}
\value{
object of class colors
}
\description{
Display colors in console
}
