% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{size}
\alias{size}
\title{Memory size of workspace objects}
\usage{
size(x, units = NULL)
}
\arguments{
\item{units}{}
}
\value{

}
\description{
Memory size of workspace objects
}
\examples{
size(ls())
}
