% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{sourcelib}
\alias{sourcelib}
\title{Source R functions from dir into local env}
\usage{
sourcelib(dir = "lib", pattern = "\\\\.r$|\\\\.R$")
}
\arguments{
\item{pattern}{}
}
\description{
Source R functions from dir into local env
}
\examples{
newenv <- sourcelib("lib")
attach(newenv)
}
