% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{stratify}
\alias{stratify}
\title{Stratify variables based on quantiles}
\usage{
stratify(x, n = 2, ties = "lower", optimal = FALSE, surv = NULL, ...)
}
\arguments{
\item{x}{Data vector}

\item{n}{Number of strata}

\item{ties}{Assign quantile ties to higher or lower group}
}
\description{
Stratify variables based on quantiles
}
