% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{summarise_cols}
\alias{summarise_cols}
\title{Summarise multiple columns by grouping}
\usage{
summarise_cols(data, coldata = NULL, by = NULL, FUN = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Summarise multiple columns by grouping
}
