% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{summarise_rows}
\alias{summarise_rows}
\title{Summarise multiple rows by grouping}
\usage{
summarise_rows(data, rowdata = NULL, by = NULL, FUN = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Summarise multiple rows by grouping
}
