% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{theme_dev}
\alias{theme_dev}
\title{Ggplot2 theme for checking plot layout and margins}
\usage{
theme_dev(theme_orig)
}
\arguments{
\item{theme_orig}{}
}
\description{
Ggplot2 theme for checking plot layout and margins
}
