% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{uniqueSets}
\alias{uniqueSets}
\title{Get unique sets (irrespective of item order)}
\usage{
uniqueSets(sets, sep = "&")
}
\arguments{
\item{sets}{list of sets}

\item{sep}{collapse set names}
}
\value{

}
\description{
Get unique sets (irrespective of item order)
}
