% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{ggseadot}
\alias{ggseadot}
\title{GSEA dotplot}
\usage{
ggseadot(
  data,
  x = -log10(padj),
  y = term,
  color = NES,
  n = 30,
  title = NULL,
  fontsize = 16,
  x_thres = 0.05,
  cex_labels = 0.6,
  lwd = 0.4,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GSEA dotplot
}
