% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Differential expression analysis with DESeq2}
\usage{
runDESeq2(
  data,
  design = NULL,
  formula = ~1,
  contrasts = NULL,
  filter = NULL,
  ctrlgenes = NULL,
  sizefactors = NULL,
  alpha = 0.05,
  ordered = TRUE,
  df = TRUE,
  ncores = NULL,
  shrink = TRUE,
  ihw = TRUE,
  vst = TRUE,
  rlog = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Differential expression analysis with DESeq2
}
