% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Differential expression analysis with DESeq2}
\usage{
runDESeq2(
  data,
  design = NULL,
  formula = ~1,
  contrasts = NULL,
  prefilter = NULL,
  postfilter = NULL,
  min_counts = 5,
  min_samples = 2,
  ctrlgenes = NULL,
  sizefactors = NULL,
  RUV = list(),
  SVA = list(),
  alpha = 0.05,
  ordered = TRUE,
  df = TRUE,
  ncores = NULL,
  shrink = TRUE,
  ihw = TRUE,
  vst = FALSE,
  rlog = FALSE,
  ...
)
}
\arguments{
\item{data}{Counts matrix}

\item{design}{Experimental design/colData}

\item{formula}{Formula}

\item{contrasts}{Named list of contrasts, specified as c(factor, level, reflevel)}

\item{prefilter}{Filter before normalization}

\item{postfilter}{Filter after normalization}

\item{min_counts}{min_counts in min_samples for filtering}

\item{min_samples}{min_counts in min_samples for filtering}

\item{ctrlgenes}{Control genes for normalization (housekeeping genes)}

\item{sizefactors}{Pre-calculated size factors}

\item{RUV}{RUV batch effect correction: 'list(empirical = \emph{genes}, group = \emph{column}, n = \emph{n_vars})'}

\item{SVA}{SVA batch effect correction: 'list(reduced = \emph{formula}, n = \emph{n_vars})'}

\item{alpha}{Significance level (default = 0.05)}

\item{ordered}{Order results (default = TRUE)}

\item{df}{Return results as data.frame}

\item{ncores}{Parallel processing}

\item{shrink}{Use ashr for log2FC shrinkage}

\item{ihw}{Use independent hypothesis weighting}

\item{vst}{Add vst-transformed assay}

\item{rlog}{Add rlog-transformed assay}

\item{...}{Parameters passed to DESeq}
}
\description{
Differential expression analysis with DESeq2
}
\examples{
dds <- DESeq2::makeExampleDESeqDataSet(n = 10000, interceptMean = c(2,5))
dds |> runDESeq2(formula = ~ condition, contrasts = list(BvsA = c("condition", "B", "A")))
dds |> runDESeq2(formula = ~ condition, contrasts = list(BvsA = c("condition", "B", "A")), vst = TRUE, ncores = 5)
dds |> runDESeq2(formula = ~ condition, RUV = list(empirical = sample(rownames(dds), 10)), contrasts = list(BvsA = c("condition", "B", "A")))
dds |> runDESeq2(formula = ~ condition, SVA = list(reduced = ~ 1), contrasts = list(BvsA = c("condition", "B", "A")))
}
