% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runGSEA}
\alias{runGSEA}
\title{GSEA with clusterProfiler}
\usage{
runGSEA(
  data,
  genesets = NULL,
  rank_by = "stat",
  rank_type = "",
  id_col = NULL,
  direction_by = NULL,
  min_size = 3,
  max_size = 2000,
  seed = 0,
  as.df = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GSEA with clusterProfiler
}
\examples{
data.frame(id = rgenes(n = 100), stat = rnorm(100)) |> col2rownames() |> runGSEA(genesets = getMSigDB(collections = "H"))
}
