% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runLIMMA}
\alias{runLIMMA}
\title{Differential expression analysis with LIMMA}
\usage{
runLIMMA(
  data,
  design,
  formula = ~1,
  contrasts = NULL,
  trend = TRUE,
  robust = FALSE,
  p.adj.method = "fdr",
  do_log = FALSE,
  normalize = FALSE,
  norm.method = "vsn",
  ...
)
}
\arguments{
\item{data}{Data}

\item{design}{Experimental design/colData}

\item{formula}{Formula}

\item{contrasts}{Named list of contrasts, specified as c(factor, level, reflevel) or c(numeric)}

\item{trend}{Trend}

\item{robust}{Robust}

\item{p.adj.method}{P.adj.method}

\item{do_log}{Log-transform the data}

\item{normalize}{Normalize}

\item{norm.method}{Normalization method}

\item{...}{}
}
\description{
Differential expression analysis with LIMMA
}
\examples{
data <- getTestData()
runLIMMA(log2(data$data+1), design = data$design, formula = ~ group + cov1, contrasts = data$contrasts)
runLIMMA(data$data, do_log = TRUE, design = data$design, formula = ~ group + cov1, contrasts = list(cov1 = "cov1"))
}
