% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{runLIMMA}
\alias{runLIMMA}
\title{Differential expression analysis with LIMMA}
\usage{
runLIMMA(
  data,
  design,
  formula = ~1,
  contrasts = NULL,
  trend = TRUE,
  robust = FALSE,
  p.adj.method = "fdr",
  normalize = FALSE,
  norm.method = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Differential expression analysis with LIMMA
}
