% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{theme_basic}
\alias{theme_basic}
\title{Basic ggplot2 theme}
\usage{
theme_basic(
  fontsize = 18,
  fontfamily = "",
  lwd = NULL,
  color = "black",
  grid.color = NULL,
  title.face = "bold",
  title.cex = 1,
  axis.title.cex = 1,
  axis.text.cex = 0.8,
  grid.cex = 0.7,
  ...
)
}
\arguments{
\item{fontsize}{Base font size}

\item{fontfamily}{Base font family}

\item{lwd}{Base linewidth}

\item{color}{Base color}

\item{grid.color}{Grid color (no grid if NULL)}

\item{title.face}{Title face (plain/bold/...)}

\item{title.cex}{Title size factor}

\item{axis.title.cex}{Axis title size factor}

\item{axis.text.cex}{Axis text size factor}

\item{grid.cex}{Grid linewidth size factor}

\item{...}{other parameters passed to theme(...)}
}
\description{
Basic ggplot2 theme
}
\examples{
ggplot(iris, aes(Sepal.Length, Petal.Length)) + geom_point() + ggtitle("Iris") + theme_basic()
ggplot(iris, aes(Sepal.Length, Petal.Length, color = Species)) + geom_point() + ggtitle("Iris") + theme_basic(grid.color = "grey70")
}
