% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{theme_basic}
\alias{theme_basic}
\title{Basic ggplot2 theme}
\usage{
theme_basic(
  base_size = 18,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  base_color = "black",
  grid = FALSE,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{...}{other parameters passed to theme(...)}
}
\description{
Basic ggplot2 theme
}
