% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{writeTables}
\alias{writeTables}
\title{Write dataframes to .xlsx file}
\usage{
writeTables(data, file, rowNames = TRUE, adjwidths = TRUE, ...)
}
\arguments{
\item{adjwidths}{}
}
\description{
Write a dataframe or a named list of dataframes to a .xlsx file
}
\examples{
writeTables(mtcars, file = "example.xlsx")
writeTables(list(mtcars = mtcars, iris = iris), file = "example.xlsx")

}
\seealso{
openxlsx::writeData
}
